hack,Makefile,test: Support running the e2e suite in a more dev-friendly configuration.
As we continue working with the refactored e2e suite, adding any additional e2e manual install tests can be cumbersome. This is because it takes a while to deploy a full install, wait for those resources to become ready to run reporting tests, then teardown any provisioned resources created. In the case where an error has occurred, then you would need to re-run the process outline above, which is not ideal.

This PR would introduce the option to skip any tasks that uninstall the metering stack, or deletes namespaces containing the e2e label. With this change, you can re-run any failed tests with an existing namespace that already contains ready pods, which leads to quicker turnaround when adding more e2e tests.

+++ hack/e2e.sh
@@ -11,8 +11,12 @@ source "${ROOT_DIR}/hack/lib/tests.sh"
 function cleanup() {
     exit_status=$?
 
-    echo "Removing namespaces with the 'name=metering-testing-ns' label"
-    kubectl delete ns -l "name=metering-testing-ns" || true
+    if [[ "${METERING_RUN_DEV_TEST_SETUP}" = false ]]; then
+        echo "Removing namespaces with the 'name=metering-testing-ns' label"
+        kubectl delete ns -l "name=metering-testing-ns" || true
+    ***REMOVED***
+        echo "Skipping the namespace deletion"
+    ***REMOVED***
 
     echo "Exiting hack/e2e.sh"
     exit "$exit_status"
@@ -31,12 +35,15 @@ TEST_LOG_FILE="${TEST_LOG_FILE:-e2e-tests.log}"
 TEST_LOG_FILE_PATH="${TEST_LOG_FILE_PATH:-$TEST_OUTPUT_DIR/$TEST_LOG_FILE}"
 TEST_JUNIT_REPORT_FILE="${TEST_JUNIT_REPORT_FILE:-junit-e2e-tests.xml}"
 TEST_JUNIT_REPORT_FILE_PATH="${TEST_JUNIT_REPORT_FILE_PATH:-$TEST_OUTPUT_DIR/$TEST_JUNIT_REPORT_FILE}"
+METERING_RUN_DEV_TEST_SETUP="${METERING_RUN_DEV_TEST_SETUP:-false}"
+
 
 mkdir -p "$TEST_OUTPUT_DIR"
 
 echo "\$KUBECONFIG=$KUBECONFIG"
 echo "\$METERING_NAMESPACE=$METERING_NAMESPACE"
 echo "\$METERING_RUN_TESTS_LOCALLY=$METERING_RUN_TESTS_LOCALLY"
+echo "\$METERING_RUN_DEV_TEST_SETUP=$METERING_RUN_DEV_TEST_SETUP"
 echo "\$METERING_OPERATOR_IMAGE_REPO=$METERING_OPERATOR_IMAGE_REPO"
 echo "\$REPORTING_OPERATOR_IMAGE_REPO=$REPORTING_OPERATOR_IMAGE_REPO"
 echo "\$METERING_OPERATOR_IMAGE_TAG=$METERING_OPERATOR_IMAGE_TAG"
@@ -60,6 +67,7 @@ go test \
     -log-level="${TEST_LOG_LEVEL}" \
     -run-tests-local="${METERING_RUN_TESTS_LOCALLY}" \
     -repo-path="${METERING_REPO_PATH}" \
+    -run-dev-setup="${METERING_RUN_DEV_TEST_SETUP}" \
     2>&1 | tee "$TEST_LOG_FILE_PATH" ; TEST_EXIT_CODE=${PIPESTATUS[0]}
 
 # if go-junit-report is installed, create a junit report also
